+++ panda_gym/envs/tasks/pick_and_place.py
@@ -93,4 +93,4 @@ class PickAndPlace(Task):
         if self.reward_type == "sparse":
             return -np.array(d > self.distance_threshold, dtype=np.float32)
         else:
-            return -d
+            return -d.astype(np.float32)
