+++ panda_gym/envs/tasks/reach.py
@@ -62,4 +62,4 @@ class Reach(Task):
         if self.reward_type == "sparse":
             return -np.array(d > self.distance_threshold, dtype=np.float32)
         else:
-            return -d
+            return -d.astype(np.float32)
